requirements.txt
requests
beautifulsoup4
rfeed
python-dateutil
lxml
